Move some thrust functions to core
Split in packages.
