+++ src/poliastro/twobody/thrust/change_inc_ecc.py
@@ -9,39 +9,9 @@ import numpy as np
 
 from astropy import units as u
 
-from poliastro.twobody.decorators import state_from_vector
-from poliastro.util import norm_fast, circular_velocity_fast
-from poliastro.twobody import rv
-from poliastro.core.jit import jit
-
-
-@jit
-def beta(ecc_0, ecc_f, inc_0, inc_f, argp):
-    # Note: "The argument of perigee will vary during the orbit transfer
-    # due to the natural drift and because e may approach zero.
-    # However, [the equation] still gives a good estimate of the desired
-    # thrust angle."
-    return np.arctan(abs(3 * np.pi * (inc_f - inc_0) / (4 * np.cos(argp) * (ecc_0 - ecc_f + np.log(
-        (1 + ecc_f) * (-1 + ecc_0) / ((1 + ecc_0) * (-1 + ecc_f)))))))
-
-
-@jit
-def delta_V(V_0, ecc_0, ecc_f, beta_):
-    """Compute required increment of velocity.
-    """
-    return 2 * V_0 * np.abs(np.arcsin(ecc_0) - np.arcsin(ecc_f)) / (3 * np.cos(beta_))
-
-
-@jit
-def extra_quantities(k, a, ecc_0, ecc_f, inc_0, inc_f, argp, f):
-    """Extra quantities given by the model.
-    """
-    beta_ = beta(ecc_0, ecc_f, inc_0, inc_f, argp)
-    V_0 = circular_velocity_fast(k, a)
-    delta_V_ = delta_V(V_0, ecc_0, ecc_f, beta_)
-    t_f_ = delta_V_ / f
-
-    return delta_V_, beta_, t_f_
+from poliastro.twobody.rv import rv2coe
+from poliastro.core.util import norm, cross
+from poliastro.core.thrust.change_inc_ecc import beta, extra_quantities
 
 
 def change_inc_ecc(ss_0, ecc_f, inc_f, f):
@@ -66,10 +36,10 @@ def change_inc_ecc(ss_0, ecc_f, inc_f, f):
     if ecc_0 > 0.001:  # Arbitrary tolerance
         ref_vec = ss_0.e_vec / ecc_0
     else:
-        ref_vec = ss_0.r / norm_fast(ss_0.r)
+        ref_vec = ss_0.r / norm(ss_0.r)
 
-    h_unit = ss_0.h_vec / norm_fast(ss_0.h_vec)
-    thrust_unit = np.cross(h_unit, ref_vec) * np.sign(ecc_f - ecc_0)
+    h_unit = ss_0.h_vec / norm(ss_0.h_vec)
+    thrust_unit = cross(h_unit, ref_vec) * np.sign(ecc_f - ecc_0)
 
     inc_0 = ss_0.inc.to(u.rad).value
     argp = ss_0.argp.to(u.rad).value
@@ -79,16 +49,16 @@ def change_inc_ecc(ss_0, ecc_f, inc_f, f):
     def a_d(t0, u_, k):
         r = u_[:3]
         v = u_[3:]
-        nu = rv.rv2coe(k, r, v)[-1]
+        nu = rv2coe(k, r, v)[-1]
         beta_ = beta_0_ * np.sign(np.cos(nu))  # The sign of ß reverses at minor axis crossings
 
-        w_ = np.cross(r, v) / norm_fast(np.cross(r, v))
+        w_ = cross(r, v) / norm(cross(r, v))
         accel_v = f * (
             np.cos(beta_) * thrust_unit +
             np.sin(beta_) * w_
         )
         return accel_v
 
-    delta_V, beta_, t_f = extra_quantities(ss_0.attractor.k.to(u.km**3 / u.s**2).value, ss_0.a.to(u.km).value,
+    delta_V, beta_, t_f = extra_quantities(ss_0.attractor.k.to(u.km ** 3 / u.s ** 2).value, ss_0.a.to(u.km).value,
                                            ecc_0, ecc_f, inc_0, inc_f, argp, f)
     return a_d, delta_V, beta_, t_f
